@@ -1,4 +1,4 @@
-name: flutter_application_09
+name: Pharera
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -31,7 +31,7 @@ dependencies:
   flutter:
     sdk: flutter
   url_launcher: ^6.0.13
-  webview_flutter: ^2.0.0
+  webview_flutter: ^2.2.0
   swipeable_button_view: ^0.0.2
   page_transition: ^2.0.9
   gradient_slide_to_act: ^0.0.2
@@ -40,34 +40,40 @@ dependencies:
   video_player_android: ^2.4.14
   chewie: ^1.0.0
   flutter_downloader: ^1.6.0
-  http: ^0.13.3
+  http: ^1.2.1
   path_provider: ^2.1.3
-  permission_handler: ^10.0.0
+  permission_handler: ^11.3.1
   curved_navigation_bar: ^1.0.3
   cupertino_icons: ^1.0.2
-  panorama_viewer: ^1.0.3
+  panorama_viewer: ^1.0.5
   image_picker: ^1.1.1
   convex_bottom_bar: ^3.2.0
   wtf_sliding_sheet: ^1.1.0
+  audioplayers: 6.0.0
+  firebase_core: ^2.31.0
+  firebase_auth: ^4.19.5
+  firebase_storage: ^11.7.5
+  fluttertoast: ^8.2.5
+  font_awesome_flutter: ^10.7.0
   flutter_localizations:
     sdk: flutter
 
-flutter_localizations:
-  sdk: flutter
-  intl: ^0.17.0
-
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  curved_navigation_bar: ^1.0.3
-  animated_bottom_navigation_bar: ^1.3.3
-  rive: ^0.13.2
-  animated_bottom_navigation: ^0.0.5
-  convex_bottom_bar: ^3.2.0
+#flutter_localizations:
+#  sdk: flutter
+#  intl: ^0.17.0
+#
+#
+#
+#  # The following adds the Cupertino Icons font to your application.
+#  # Use with the CupertinoIcons class for iOS style icons.
+#  cupertino_icons: ^1.0.2
+#  curved_navigation_bar: ^1.0.3
+#  animated_bottom_navigation_bar: ^1.3.3
+#  rive: ^0.13.2
+#  animated_bottom_navigation: ^0.0.5
+#  convex_bottom_bar: ^3.2.0
 dev_dependencies:
-  flutter_gen: ^4.0.0
+  flutter_gen: ^5.5.0+1
   flutter_test:
     sdk: flutter
 
@@ -76,7 +82,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -92,10 +98,24 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
-    - assets/videos/
     - assets/images/360_images/
-    - assets/videos/Libsync/
+    - assets/images/tour_guide/
     - assets/images/artifacts/
+    - assets/images/icons/
+    - assets/images/exi/
+
+    - assets/videos/
+    - assets/videos/Libsync/
+    - assets/videos/tour_guide/
+
+    - assets/subtitles/
+    - assets/sub_libsync/
+    - assets/sub_libsync/0/
+    - assets/sub_libsync/1/
+    - assets/sub_libsync/2/
+    - assets/sub_libsync/3/
+    - assets/sub_libsync/4/
+    - assets/sub_libsync/5/
 #    - assets/RiveAssets/
     
 
