@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:Pharera/notifacation.dart';
-import 'package:Pharera/pharahosl_ist.dart';
+import 'package:Pharera/pharahos_list.dart';
 import 'package:Pharera/pharaoh_show.dart';
 import 'package:Pharera/navigation_bar.dart';
 import 'package:page_transition/page_transition.dart';
 import 'package:wtf_sliding_sheet/wtf_sliding_sheet.dart';
 
 class Groub extends StatefulWidget {
-  const Groub({Key? key}) : super(key: key);
+  const Groub({super.key});
 
   @override
   _GroubState createState() => _GroubState();
@@ -20,6 +20,7 @@ class _GroubState extends State<Groub> {
   String _selectedDetails = "";
   int _selectedIndex = 0;
   final SheetController _sheetController = SheetController();
+  bool _isSheetAttached = false;
 
   @override
   void initState() {
@@ -42,7 +43,7 @@ class _GroubState extends State<Groub> {
                 MaterialPageRoute(builder: (context) => const Not()),
               );
             },
-            child: Container(
+            child: SizedBox(
               width: 40,
               height: 40,
               child: Image.asset('assets/images/notification.jpg'),
@@ -79,99 +80,114 @@ class _GroubState extends State<Groub> {
                   ),
                 ),
               ),
-              const SizedBox(height: 64),
+              const SizedBox(height: 10),
               _selectedPhoto.isNotEmpty
                   ? Container(
-                width: 340,
-                alignment: AlignmentDirectional.bottomCenter,
-                height: 400,
-                decoration: BoxDecoration(
-                  color: Colors.transparent,
-                  borderRadius: const BorderRadius.all(Radius.circular(24)),
-                  image: DecorationImage(
-                    fit: BoxFit.fill,
-                    image: AssetImage(
-                      pharaohData.pharaoh[_selectedIndex]['image'] ?? "",
-                    ),
-                  ),
-                ),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  children: [
-                    Text(
-                      _selectedName,
-                      style: const TextStyle(color: Colors.transparent, fontSize: 24),
-                    ),
-                    const SizedBox(height: 24),
-                    SizedBox(
-                      width: 110,
-                      height: 48,
-                      child: ElevatedButton(
-                        onPressed: () {
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                              builder: (context) => PharaohDetailPage(
-                                pharaohData: pharaohData,
-                                index: _selectedIndex,
-                              ),
-                            ),
-                          );
-                        },
-                        child: const Text(
-                          "EXPLORE",
-                          style: TextStyle(color: Colors.black),
+                      width: 340,
+                      alignment: AlignmentDirectional.bottomCenter,
+                      height: 400,
+                      decoration: BoxDecoration(
+                        color: Colors.transparent,
+                        borderRadius:
+                            const BorderRadius.all(Radius.circular(24)),
+                        image: DecorationImage(
+                          fit: BoxFit.fill,
+                          image: AssetImage(
+                            pharaohData.pharaoh[_selectedIndex]['image'] ?? "",
+                          ),
                         ),
                       ),
+                      child: Column(
+                        mainAxisAlignment: MainAxisAlignment.end,
+                        children: [
+                          Text(
+                            _selectedName,
+                            style: const TextStyle(
+                                color: Colors.transparent, fontSize: 24),
+                          ),
+                          const SizedBox(height: 24),
+                          SizedBox(
+                            width: 110,
+                            height: 48,
+                            child: ElevatedButton(
+                              onPressed: () {
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                    builder: (context) => PharaohDetailPage(
+                                      pharaohData: pharaohData,
+                                      index: _selectedIndex,
+                                    ),
+                                  ),
+                                );
+                              },
+                              child: const Text(
+                                "EXPLORE",
+                                style: TextStyle(color: Colors.black),
+                              ),
+                            ),
+                          )
+                        ],
+                      ),
                     )
-                  ],
-                ),
-              )
                   : Container(
-                width: 340,
-                height: 540,
-                decoration: BoxDecoration(
-                  color: Colors.transparent,
-                  borderRadius: const BorderRadius.all(Radius.circular(24)),
-                  image: DecorationImage(
-                    fit: BoxFit.fill,
-                    image: AssetImage(pharaohData.pharaoh.firstWhere((e) => e['name'] == "King Akhenaten")['image'] ?? ""),
-                  ),
-                ),
-              ),
+                      width: 340,
+                      height: 540,
+                      decoration: BoxDecoration(
+                        color: Colors.transparent,
+                        borderRadius:
+                            const BorderRadius.all(Radius.circular(24)),
+                        image: DecorationImage(
+                          fit: BoxFit.fill,
+                          image: AssetImage(pharaohData.pharaoh.firstWhere(
+                                  (e) =>
+                                      e['name'] == "King Akhenaten")['image'] ??
+                              ""),
+                        ),
+                      ),
+                    ),
             ],
           ),
-          SlidingSheet(
-            controller: _sheetController,
-            elevation: 8,
-            cornerRadius: 16,
-            snapSpec: const SnapSpec(
-              snap: true,
-              snappings: [0.3, 0.6, 1.0],
-              positioning: SnapPositioning.relativeToAvailableSpace,
-            ),
-            builder: (context, state) {
-              return Container(
-                height: MediaQuery.of(context).size.height * 0.7,
-                child: Padding(
-                  padding: const EdgeInsets.all(16.0),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        _selectedName,
-                        style: const TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        _selectedDetails,
-                        style: const TextStyle(fontSize: 16),
-                      ),
-                    ],
-                  ),
-                ),
-              );
+          GestureDetector(
+            onTap: () {
+              if (!_isSheetAttached) {
+                _sheetController.snapToExtent(0.7);
+                _isSheetAttached = true;
+              }
             },
+            child: SlidingSheet(
+              controller: _sheetController,
+              elevation: 8,
+              cornerRadius: 16,
+              snapSpec: const SnapSpec(
+                snap: true,
+                snappings: [0.25, 0.6, 1.0],
+                positioning: SnapPositioning.relativeToAvailableSpace,
+              ),
+              builder: (context, state) {
+                return Container(
+                  height: MediaQuery.of(context).size.height * 0.7,
+                  child: Padding(
+                    padding: const EdgeInsets.all(16.0),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          _selectedName,
+                          style: const TextStyle(
+                              fontSize: 24, fontWeight: FontWeight.bold),
+                        ),
+                        const SizedBox(height: 16),
+                        Text(
+                          _selectedDetails,
+                          style: const TextStyle(fontSize: 16),
+                        ),
+                      ],
+                    ),
+                  ),
+                );
+              },
+            ),
           ),
         ],
       ),
@@ -187,7 +203,7 @@ class _GroubState extends State<Groub> {
           _selectedDetails = pharaohData.pharaoh[index]['details']!;
           _selectedIndex = index;
         });
-        _sheetController.snapToExtent(0.7);
+        _isSheetAttached = false; // Reset the attachment flag
       },
       child: Container(
         width: 50,
@@ -195,7 +211,9 @@ class _GroubState extends State<Groub> {
         decoration: BoxDecoration(
           shape: BoxShape.circle,
           border: Border.all(
-            color: _selectedPhoto == pharaohData.pharaoh[index]['name'] ? Colors.yellow : Colors.transparent,
+            color: _selectedPhoto == pharaohData.pharaoh[index]['name']
+                ? Colors.yellow
+                : Colors.transparent,
             width: 2,
           ),
           image: DecorationImage(
