@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_09/Sign_up.dart';
-import 'package:flutter_application_09/sign.dart';
-
+// import 'package:Pharera/sign_up.dart';
+// import 'package:Pharera/sign_in.dart';
+import 'package:Pharera/features/user_auth/presentation/pages/login_page.dart';
+import 'package:Pharera/features/user_auth/presentation/pages/sign_up_page.dart';
 class Register extends StatelessWidget {
-  const Register({Key? key});
+  const Register({super.key, Key});
 
   @override
   Widget build(BuildContext context) {
@@ -26,7 +27,7 @@ class Register extends StatelessWidget {
                   onTap: () {
                     Navigator.push(
                       context,
-                      MaterialPageRoute(builder: (context) => const SignIn()),
+                      MaterialPageRoute(builder: (context) => const LoginPage()),
                     );
                   },
                   child: Container(
@@ -34,7 +35,19 @@ class Register extends StatelessWidget {
                     height: 80,
                     decoration: BoxDecoration(
                       borderRadius: BorderRadius.circular(20),
-                      color: const Color(0xFFD5B774),
+                      gradient: const LinearGradient(
+                        colors: [
+                          Color(0xFFCCAE6B),
+                          Color(0xFFD5B975),
+                          Color(0xFFD4B97C),
+                          Color(0xFFE0C586),
+                          Color(0xFFEED392),
+                          Color(0xFFE8CD8C),
+                          Color(0xFFDABF80),
+                          Color(0xFFCBB073),
+                          Color(0xFFBBA267),
+                        ],
+                      ),
                       boxShadow: [
                         BoxShadow(
                           color: Colors.black.withOpacity(0.5),
@@ -63,7 +76,7 @@ class Register extends StatelessWidget {
                   onTap: () {
                     Navigator.push(
                       context,
-                      MaterialPageRoute(builder: (context) => const Signuo()),
+                      MaterialPageRoute(builder: (context) => const SignUpPage()),
                     );
                   },
                   child: Container(
@@ -71,7 +84,19 @@ class Register extends StatelessWidget {
                     height: 80,
                     decoration: BoxDecoration(
                       borderRadius: BorderRadius.circular(20),
-                      color: const Color(0xFFD5B774),
+                      gradient: const LinearGradient(
+                        colors: [
+                          Color(0xFFCCAE6B),
+                          Color(0xFFD5B975),
+                          Color(0xFFD4B97C),
+                          Color(0xFFE0C586),
+                          Color(0xFFEED392),
+                          Color(0xFFE8CD8C),
+                          Color(0xFFDABF80),
+                          Color(0xFFCBB073),
+                          Color(0xFFBBA267),
+                        ],
+                      ),
                       boxShadow: [
                         BoxShadow(
                           color: Colors.black.withOpacity(0.5),
