@@ -1,8 +1,195 @@
 class PharaohData {
   List<Map<String, String>> pharaoh = [
+
+    {
+      "image": "assets/images/amnthotb el talt (1).jpg",
+      "name": "King Amnthotb III",
+      "namee": "امنحتب الثالث",
+      "details": "Amenhotep III, also known as Amenhotep the Magnificent, ruled ancient Egypt during the 18th dynasty from around 1386 to 1353 BCE. His reign was marked by prosperity, cultural flourishing, and extensive building projects, including the Temple of Luxor and the Colossi of Memnon. He was known for his diplomatic prowess, maintaining peaceful relations and trade with neighboring kingdoms.\n\n"
+          "Amenhotep III, also known as Amenhotep the Magnificent, ruled over Egypt during the 18th dynasty (circa 1386–1353 BCE) and achieved numerous accomplishments during his reign. Here are some of his notable achievements:\n\n"
+          "1. Building Projects: Amenhotep III was renowned for his extensive building projects, including the construction and renovation of temples, palaces, and monuments throughout Egypt. He built or expanded temples at Karnak, Luxor, and Soleb, among others. His mortuary temple, known as the 'Temple of Millions of Years,' was located on the west bank of the Nile at Thebes and was one of the largest and most impressive structures of its time.\n\n"
+          "2. Art and Culture: Amenhotep III was a patron of the arts and presided over a flourishing period of artistic and cultural achievement in ancient Egypt. He commissioned numerous statues, reliefs, and other works of art, many of which depicted him and his family in idealized forms. The period of his reign is often considered a golden age of Egyptian art.\n\n"
+          "3. Foreign Relations: Amenhotep III maintained peaceful relations with foreign powers during his reign, which contributed to stability and prosperity within Egypt. He engaged in diplomatic marriages with foreign princesses to strengthen alliances, including his marriage to Tiy, who became one of the most influential queens in Egyptian history.\n\n"
+          "4. Economic Prosperity: Under Amenhotep III's rule, Egypt experienced a period of economic prosperity and abundance. Trade flourished, bringing wealth and exotic goods from distant lands, and the agricultural sector thrived, leading to increased food production and prosperity for the population.\n\n"
+          "5. Religious Contributions: Amenhotep III was deeply involved in the religious life of Egypt and sponsored the construction and renovation of temples dedicated to various gods and goddesses. He was particularly devoted to the god Amun-Ra, the principal deity of the Egyptian pantheon, and his religious activities contributed to the power and influence of the priesthood of Amun.\n\n"
+          "6. Monumental Statues: Amenhotep III commissioned numerous colossal statues of himself, including the famous Colossi of Memnon, which stood at the entrance to his mortuary temple. These statues, carved from massive blocks of stone, were intended to showcase the pharaoh's power and divine status.\n\n"
+          "7. Longevity and Stability: Amenhotep III's reign was one of the longest and most stable in ancient Egyptian history, lasting for nearly four decades. His ability to maintain peace, stability, and prosperity throughout his reign contributed to his reputation as one of Egypt's most successful and beloved pharaohs."
+    },
+
+    {"image": "assets/images/hatshbsot.jpg", "name": "King Hatshbsot",
+      "namee": "حستبشوت",
+      "details":"Hatshepsut, one of ancient Egypt's most renowned female rulers, reigned as pharaoh from around 1479 to 1458 BCE during the 18th dynasty. She famously dressed as a male pharaoh to assert her authority and legitimacy, commissioning ambitious building projects such as the Temple of Hatshepsut at Deir el-Bahari. Under her rule, Egypt experienced peace, prosperity, and extensive trade expeditions to the Land of Punt, enriching the kingdom's resources and cultural exchange\n\n"
+          "Hatshepsut, one of ancient Egypt's most remarkable rulers, achieved numerous accomplishments during her reign as pharaoh (circa 1478–1458 BCE). Here are some of her notable achievements:\n\n"
+          "1. First Female Pharaoh: Hatshepsut is perhaps best known for being one of the few women to rule as pharaoh in ancient Egypt. She took the throne after the death of her husband, Thutmose II, and acted as regent for her stepson, Thutmose III, before eventually assuming full pharaonic powers.\n\n"
+          "2. Building Projects: She initiated ambitious building projects throughout Egypt, including the construction of magnificent temples and monuments. The most famous of these is her mortuary temple, Djeser-Djeseru (\"Holy of Holies\"), located at Deir el-Bahri near Thebes. This temple is renowned for its innovative architecture and stunning reliefs.\n\n"
+          "3. Trade and Diplomacy: Hatshepsut's reign saw an expansion of trade networks and diplomatic relations with other civilizations, particularly in the eastern Mediterranean and the land of Punt (possibly modern-day Somalia or Eritrea). Trade missions to Punt brought back exotic goods such as myrrh, ebony, ivory, and live animals.\n\n"
+          "4. Religious Contributions: Hatshepsut played a significant role in the religious life of Egypt, promoting the worship of traditional gods and goddesses while also elevating the status of some lesser-known deities. She associated herself closely with the god Amun, claiming divine lineage and portraying herself as his earthly representative.\n\n"
+          "5. Cultural Patronage: Hatshepsut supported art, literature, and culture during her reign, commissioning statues, reliefs, and other works of art that celebrated her accomplishments and perpetuated her legacy.\n\n"
+          "6. Stability and Prosperity: Under Hatshepsut's rule, Egypt experienced a period of stability and prosperity, with economic growth, trade expansion, and cultural flourishing. Her reign is often remembered as a time of peace and prosperity in ancient Egyptian history.\n\n"
+          "7. Legacy and Succession: Despite efforts to erase her legacy by later pharaohs, including her stepson Thutmose III, Hatshepsut's achievements endured through the centuries. She remains one of ancient Egypt's most celebrated and enigmatic rulers, remembered for her exceptional leadership and the lasting impact of her reign."},
+
+    {"image": "assets/images/akhnton.jpg", "name": "King Akhenaten",
+      "namee": "اخناتون",
+      "details":"Akhenaten, known for his revolutionary religious reforms, ruled ancient Egypt during the 18th dynasty from around 1353 to 1336 BCE. He introduced monotheistic worship centered on the sun god Aten, shifting away from traditional polytheism. His capital, Akhetaten (modern-day Amarna), reflected his new religious ideology. Despite initial prosperity, his reign faced opposition, and after his death, his reforms were largely reversed, and his monuments were dismantled.\n\n"
+          "Akhenaten, also known as Amenhotep IV, is one of ancient Egypt's most intriguing and controversial pharaohs, primarily known for his religious reforms and radical changes in artistic style. Here are some of his notable achievements:\n\n"
+          "1. Religious Revolution: Akhenaten is most famous for his religious revolution, which involved the elevation of the sun god Aten to the status of the supreme deity and the suppression of traditional Egyptian polytheism. He proclaimed Aten as the one true god and initiated a monotheistic cult centered around the worship of the sun disc. This represented a significant departure from the traditional polytheistic beliefs of ancient Egypt.\n\n"
+          "2. Establishment of Akhetaten: Akhenaten founded a new capital city called Akhetaten (modern-day Amarna) dedicated to the worship of Aten. The city was constructed from scratch and served as the administrative and religious center of Egypt during Akhenaten's reign. It featured grand temples and palaces, reflecting the new religious and artistic ideals of the Aten cult.\n\n"
+          "3. Artistic Innovations: Akhenaten's reign witnessed a revolution in artistic style known as the \"Amarna art.\" This art style deviated from traditional Egyptian artistic conventions, depicting the pharaoh and his family in a more naturalistic and intimate manner. It emphasized realism, with elongated proportions, exaggerated features, and a focus on the royal family's everyday activities.\n\n"
+          "4. Royal Family: Akhenaten's family played a prominent role in his religious reforms and artistic innovations. His wife, Nefertiti, was a powerful and influential queen who played a central role in the Aten cult. Their daughters, especially the famous Ankhesenamun, were also prominent figures in Akhenaten's court.\n\n"
+          "5. Diplomacy and Foreign Policy: Akhenaten's reign saw a shift in Egypt's foreign policy, with a focus on diplomacy and pacifism rather than military conquest. He sought to maintain peaceful relations with neighboring powers and engaged in diplomatic correspondence with other rulers, including those in the Hittite Empire.\n\n"
+          "6. Literary Contributions: Akhenaten is credited with composing a series of hymns and prayers dedicated to the Aten, known as the \"Great Hymn to the Aten.\" These texts praise the sun god and express Akhenaten's devotion to Aten and his role as the sole mediator between Aten and humanity.\n\n"
+          "7. Legacy and Controversy: Despite his significant achievements, Akhenaten's reign ended in controversy, and his religious reforms were largely dismantled by his successors. Nevertheless, his reign represents a fascinating period of experimentation and innovation in ancient Egyptian history, leaving a lasting impact on art, religion, and culture in the Nile Valley."},
+
+    {"image": "assets/images/sonsrt3.jpg", "name": "King Sonsrt III",
+      "namee": "سنوسيرت الثالث",
+      "details":"Senusret III, a pharaoh of the Middle Kingdom's 12th dynasty, ruled ancient Egypt from approximately 1878 to 1839 BCE. He is celebrated for his military campaigns, expanding Egypt's borders and securing its territories. Senusret III initiated numerous construction projects, including the famous White Chapel at Karnak. His reign is regarded as a period of stability and prosperity, often considered one of the peaks of Egyptian civilization.\n\n"
+          "Senusret III, also known as Sesostris III, was a powerful pharaoh of the Middle Kingdom of Egypt who ruled during the 12th Dynasty (circa 1878–1839 BCE). He is often regarded as one of the most capable and successful rulers in ancient Egyptian history. Here are some of his notable achievements:\n\n"
+          "1. Military Campaigns: Senusret III conducted numerous military campaigns to expand Egypt's territory and maintain control over its borders. He waged successful campaigns against Nubia (to the south), the Levant (to the northeast), and Libya (to the west), securing Egypt's dominance in the region.\n\n"
+          "2. Fortification Projects: Senusret III undertook extensive fortification projects along Egypt's borders to protect the kingdom from external threats. He built a series of fortresses, forts, and military outposts along strategic points on the Nile and in key regions such as the Sinai Peninsula and Lower Nubia.\n\n"
+          "3. Infrastructure Development: Senusret III initiated significant infrastructure projects aimed at improving transportation, communication, and irrigation systems throughout Egypt. He constructed or renovated canals, dikes, and waterways to control the annual flooding of the Nile and ensure agricultural prosperity.\n\n"
+          "4. Monumental Architecture: Senusret III commissioned several impressive architectural projects, including temples, palaces, and statues dedicated to the gods and to commemorate his reign. Notable examples include the renovation of the Temple of Karnak in Thebes and the construction of the White Chapel, a small temple made of white limestone adorned with reliefs depicting the king's divine status.\n\n"
+          "5. Administrative Reforms: Senusret III implemented administrative reforms aimed at centralizing royal authority and strengthening the bureaucratic apparatus of the Egyptian state. He appointed trusted officials to key administrative positions and established systems for taxation, law enforcement, and land management.\n\n"
+          "6. Cultural Patronage: Senusret III was a patron of the arts and literature, supporting the production of religious texts, hymns, and literary works that celebrated his achievements and glorified the gods. His reign witnessed a flourishing of intellectual and artistic activity, with scholars and artisans producing works of exceptional quality and beauty.\n\n"
+          "7. Longevity and Stability: Senusret III's reign was characterized by stability, prosperity, and relative peace, allowing Egypt to thrive economically and culturally. His efficient governance and military prowess contributed to the longevity of the 12th Dynasty and secured his legacy as one of ancient Egypt's most successful pharaohs."},
+
+    {"image": "assets/images/sonsrt1.jpg", "name": "King Sonsrt I",
+      "namee": "سنوسرت الاول",
+      "details":"Senusret I, a pharaoh of the Middle Kingdom's 12th dynasty, ruled ancient Egypt from approximately 1971 to 1926 BCE. He is renowned for his military campaigns, notably against the Nubians, expanding Egypt's territory and influence. Senusret I oversaw numerous construction projects, including the famous White Chapel and the Pyramid of El-Lahun. His reign is remembered as a period of prosperity and cultural flourishing, marking one of the heights of the Middle Kingdom.\n\n"
+          "Senusret I, also known as Senwosret I, was a significant pharaoh of the Middle Kingdom's 12th Dynasty in ancient Egypt, ruling from around 1971 to 1926 BCE. He is often regarded as one of the most capable and influential rulers of his dynasty. Here are some of his notable achievements:\n\n"
+          "1. Building Projects: Senusret I initiated several ambitious building projects throughout Egypt, including the construction and renovation of temples, palaces, and monuments. He built temples dedicated to various gods and goddesses, such as the Temple of Amun at Karnak and the Temple of Montu at Tod.\n\n"
+          "2. Expansion of Irrigation: Senusret I is credited with expanding Egypt's irrigation systems, particularly in the Fayyum region. He commissioned the construction of canals and waterways to improve agricultural productivity and increase the arable land available for cultivation, leading to economic prosperity.\n\n"
+          "3. Military Campaigns: Senusret I conducted successful military campaigns to defend Egypt's borders and expand its territory. He waged campaigns against the Nubians in the south and the Libyans in the west, securing Egypt's dominance over these regions and ensuring stability along the borders.\n\n"
+          "4. Monumental Statuary: Senusret I commissioned numerous statues and reliefs depicting himself in various poses and royal regalia. These statues served not only to glorify the king but also to propagate royal ideology and affirm his divine kingship to the Egyptian people.\n\n"
+          "5. Cultural Patronage: Senusret I was a patron of the arts and literature, supporting the production of religious texts, hymns, and literary works that celebrated the achievements of the pharaoh and extolled the virtues of the gods. His reign witnessed a flourishing of intellectual and artistic activity, with scholars and artisans producing works of exceptional quality.\n\n"
+          "6. Strengthening of Central Authority: Senusret I implemented administrative reforms aimed at centralizing royal authority and strengthening the bureaucratic apparatus of the Egyptian state. He appointed trusted officials to key administrative positions and established systems for taxation, law enforcement, and land management.\n\n"
+          "7. Legacy and Succession: Senusret I's reign marked a period of stability and prosperity in ancient Egypt, laying the foundation for the subsequent glory of the 12th Dynasty. His efficient governance and military prowess secured his legacy as one of Egypt's most successful and revered pharaohs, revered by later generations for his accomplishments and contributions to the kingdom."},
+
+    {"image": "assets/images/to7tmos.jpg", "name": "King Tahtmos III",
+      "namee": "تحتمس الثالث",
+      "details":"Thutmose III, often referred to as the Napoleon of Egypt, reigned during the New Kingdom's 18th dynasty, from about 1479 to 1425 BCE. He expanded Egypt's empire through numerous military campaigns, including the famous Battle of Megiddo. Thutmose III implemented administrative reforms and encouraged trade, leading to a period of prosperity and cultural advancement known as the Thutmoside Renaissance His reign solidified Egypt's status as a dominant power in the ancient Near East.\n\n"
+          "Thutmose III, often referred to as the Napoleon of Egypt was one of ancient Egypt's most successful warrior pharaohs, reigning from approximately 1479 to 1425 BCE during the New Kingdom period. His achievements are numerous and significant, shaping the course of Egyptian history. Here are some of his notable accomplishments:\n\n"
+          "1. Military Campaigns and Expansion: Thutmose III conducted numerous military campaigns to expand Egypt's borders and establish its dominance in the ancient Near East. He led at least 17 military campaigns into Syria, Canaan, and Nubia, extending Egypt's territory to its greatest extent.\n\n"
+          "2. Battle of Megiddo: Thutmose III's most famous military victory occurred at the Battle of Megiddo, where he defeated a coalition of Canaanite and other regional powers. This decisive victory is one of the earliest recorded battles in history and solidified Egypt's control over the Levant.\n\n"
+          "3. Administrative Reforms: Thutmose III implemented administrative reforms aimed at centralizing royal authority and streamlining the bureaucracy. He appointed trusted officials to key administrative positions and established systems for taxation, law enforcement, and land management.\n\n"
+          "4. Monumental Building Projects: Thutmose III undertook extensive building projects, including the construction and renovation of temples, palaces, and monuments throughout Egypt. Notable examples include the expansion of the Karnak Temple complex and the construction of the Festival Hall at Karnak.\n\n"
+          "5. Cultural Patronage: Thutmose III was a patron of the arts and literature, supporting the production of religious texts, hymns, and artistic works that celebrated his military victories and glorified the gods. He commissioned statues and reliefs depicting himself in various poses and royal regalia.\n\n"
+          "6. Diplomacy and Trade: Thutmose III engaged in diplomacy and trade with foreign powers, forging alliances and establishing commercial relations with neighboring kingdoms. He negotiated treaties and marriage alliances to secure peace and strengthen Egypt's position in the region.\n\n"
+          "7. Temple Restorations: Thutmose III restored and renovated many temples across Egypt, dedicating resources to the upkeep and maintenance of religious structures. He also commissioned new temples and monuments to honor the gods and commemorate his military victories.\n\n"
+          "8. Longevity and Stability: Thutmose III's reign marked a period of stability and prosperity in ancient Egypt, known as the 'Golden Age' of the New Kingdom. His military conquests and administrative reforms contributed to Egypt's wealth and power, ensuring his legacy as one of Egypt's greatest pharaohs."},
+
+
+    /// exibition ///
+    {
+      "image": "assets/images/exi/masala.jpg",
+      "name": "The Hanging Obelisk of King Ramesses II",
+      "namee": "المسلة المعلقة للملك رمسيس الثاني",
+      "details": "Now we will talk about this exciting artifact. This obelisk was discovered in a place called Sohag El Hegaza in the Eastern region, currently divided into several parts. It is carved from pink granite and is distinguished by the beauty of its inscriptions, which depict King Ramesses II standing in front of one of the deities, in addition to the various titles bestowed upon him."
+
+    },
+
+
+    {
+      "image": "assets/images/exi/tahtmos.jpg",
+      "name": "Thutmose III",
+      "namee": "تمثال للملك تحتمس الثالث",
+      "details": "- We will know more about his history through lip syncing. - King Thutmose III is a pharaoh from the 18th dynasty of ancient Egypt. He ruled during the second intermediate period, and his history is filled with intriguing events to talk about now."
+    },
+
+    {
+      "image": "assets/images/exi/3mod.jpg",
+      "name": "victory column of king merenptah",
+      "namee": "عمود النصر للملك مرنبتاح",
+      "details": "Now we will talk about this fascinating artifact. This stele was discovered in a place called Sohag El-Hagar in the Eastern Desert. It is currently divided into several parts and is carved from pink granite. It is distinguished by the beauty of its inscriptions, which depict King Ramses II standing in front of one of the deities, in addition to his various titles."
+    },
+    {
+      "image": "assets/images/sonsrt1.jpg",
+      "name": "sunesurt I",
+      "namee": "سونسرت الاول",
+      "details": "This head belongs to an early period of the reign of King Akhenaten. It reflects the features of the royal face with its narrow almond-shaped eyes and long face, in the style of the Amarna art, which characterized that period of Egyptian history. It was found in the vicinity of the temple dedicated to the sun god (Aten)."
+    },
+    {
+      "image": "assets/images/exi/ramses.jpg",
+      "name": "Rameses II",
+      "namee": "رمسيس الثاني",
+      "details": "This huge statue shows the extent of the power and wealth of its owner, King Ramses II. This statue was originally standing outside the main temple of the god Ptah in the city of Memphis, the ancient capital of the country, where the purpose of representing the king in this majestic form was to make Ramses a guardian of the sanctuary of that sacred temple, and we will see The history of this ancient king is in its current form as it unfolds now"
+    },
+    {
+      "image": "assets/images/exi/colo.jpg",
+      "name": "colossal statue of a Middle kingdom king",
+      "namee": "تمثال لملك من الدولي الوسطي",
+      "details": "-This statue depicts one of the two kings; Senusret III, or Amenemhat IV. Two princesses appear on the side of the statue's legs. It is not yet certain for us whether King Amenemhat IV had children. As for King Senusret III, he had three wives and five daughters."
+    },
+    {
+      "image": "assets/images/sonsrt3.jpg",
+      "name": "senusret III",
+      "namee": "سونسرت الثالث",
+      "details": "He was one of the fifth pharaohs of the Twelfth Dynasty, and is considered one of the greatest pharaohs of the Middle Kingdom. The Greeks called him Sesostris III, and he will tell many stories about himself now."
+    },
+
+    {
+      "image":"assets/images/exi/abo.jpg",
+      "name": "The Sphinx of King Amenemhat III",
+      "namee": "تمثال ابو الهول للملك امنمحات الثالث",
+      "details": "The Sphinx is considered an amalgamation of the power of the lion and the power of the king at the same time. This statue is one of seven statues of the Sphinx of King Amenemhat III, which were found in Tanis in the Delta. The features of these seven statues are similar to the features of the king, as they show the king with a full face and wide lips."
+    },
+    {
+      "image": "assets/images/exi/abo2.jpg",
+      "name": "The Sphinx statue with the names of the kings Ramesses II and Merneptah",
+      "namee": "تمثال ابي الهول بأسماء الملكين رمسيس الثاني و مرنبتاح",
+      "details": "-The statue depicts the Sphinx with the body of a lion and the head of a king, and is a famous example of ancient Egyptian art that expresses royal and divine power. The statue is carved from limestone, the same material used in the Great Sphinx of Giza. The statue reflects the importance of the Sphinx in ancient Egyptian civilization as a symbol of royal protection and power."
+    },
+    {
+      "image": "assets/images/exi/32.jpg",
+      "name": "Statue of King Ramesses II with the deity Ptah Tatenen",
+      "namee": "تمثال الملك رمسيس الثاني مع المعبود بتاح تاتنن" ,
+      "details": "--God of creative creativity: Ptah Tatenen was worshiped in ancient Egypt as the god of creation and crafts. Ptah was considered one of the major gods in the city of Memphis, the scientific and administrative center of ancient Egypt. Ptah was the maker of gods and men, and said he created the world through the logic of sacred words.."
+    },
+    {
+      "image": "assets/images/exi/hat.jpg",
+      "name": "Hathorian column capital",
+      "namee": "تاج عمود حتحوري" ,
+      "details": "-This piece represented the crown of one of the Hathor-style columns, as it was depicted in This piece represented the crown of one of the Hathor-style columns, Aboudah Hathor, with a human face, one of the musical instruments known as the salas, and a cobra snake, and the complete shape of these columns resembles the ears of a cow, and a small chapel is surmounted above its head. It contains the Thirteenth Dynasty or the beginning of the Ptolemaic period, which women carried and played by shaking it with their hands in religious rituals. This crown may be due to one of the lost temples in the city of Mendes."
+    },
+    {
+      "image": "assets/images/exi/22.jpg",
+      "name": "Statue of King Ramesses II with the goddess Anat",
+      "namee": "تمثال الملك رمسيس الثاني مع المعبودة عنات" ,
+      "details": "Anat was worshiped as a goddess of war and fertility in Canaanite and ancient Egyptian mythology. She is usually depicted as a woman carrying a sword or spear and wearing a crown with feathers. She played an important role in enhancing the concept of strength and protection among kings and warriors in ancient cultures."
+    },
+    {
+      "image": "assets/images/exi/rams2.jpg",
+      "name": "A statue of King Ramesses II protecting a deity",
+      "namee": "تمثال للملك رمسيس الثاني في حماية احدي المعبودات",
+      "details": "He is the son of Pharaoh Seti I and Queen Tuya, and was called the co-ruler of his father. He accompanied his father during his military campaigns in Nubia, the Levant, and Libya when he was fourteen years old. During his reign, Ramesses II led several military campaigns to the Levant and built a huge number of temples and statues, the most famous of which is the Abu Simbel Temple in Aswan Governorate, in the south of the country. The Egyptian Pharaoh also built a large number of obelisks."
+    },
+    {
+      "image": "assets/images/exi/oz.jpg",
+      "name": "Statue of Osiris",
+      "namee": "تمثال اوزوريس" ,
+      "details": "-Osiris is the god of the afterlife, the underworld, and the dead in ancient Egyptian mythology. It is also associated with rebirth, life after death, and fertility. Osiris is considered one of the most important gods in the Egyptian pantheon, and is the focus of myths related to death and the afterlife. Statues of Osiris were commonly placed in temples, tombs, and other sacred places to honor the god and invoke his protection and blessings for the afterlife. They served a religious and protective function, ensuring that the deceased achieved eternal life."
+    },
+    {
+      "image": "assets/images/exi/3.jpg",
+      "name": "Statue of the god Ptah and King Ramesses II with the goddess Sekhmet",
+      "namee": "تمثال المعبود بتاح و الملك رمسيس الثاني مع المعبودة سخمت",
+      "details": "He is the son of Pharaoh Seti I and Queen Tuya, and was called the co-ruler of his father. He accompanied his father during his military campaigns in Nubia, the Levant, and Libya when he was fourteen years old. During his reign, Ramesses II led several military campaigns to the Levant and built a huge number of temples and statues. Sekhmet was the goddess of war and destruction but also healing and protection in ancient Egypt. Sekhmet was considered the protector of the pharaoh and played an important role in rituals related to war. Ptah was worshiped in ancient Egypt, especially in the city of Memphis (Memphis), and was considered the god of creation, industry, and the arts. It is considered a symbol of wisdom and ingenuity in industry."
+    },
+    {
+      "image": "assets/images/exi/big.jpg",
+      "name": "A huge statue of the god Ptah",
+      "namee": "تمثال ضخم للمعبود بتاح" ,
+      "details": "-The statue of the god Ptah represents the god of creation, industry, and the arts in ancient Egypt. He was worshiped mainly in the city of Memphis (Memphis) and is depicted as a man wrapped in a narrow ribbon with a scepter that brings together symbols of life, stability and strength. The statues are usually made of stone or metal and feature fine details. The statue is considered an important symbol that reflects the Egyptian prowess in sculpture and the importance of Ptah in Egyptian religion."
+    },
+    {
+      "image": "assets/images/exi/2.jpg",
+      "name": "A double statue of King Amenhotep III with the goddess Ra Hor-akheti",
+      "namee": "تمثال مزدوج للملك امنحتب الثالث مع المعبودة رع حور اختي",
+      "details": "King Amenhotep III sits in this statue next to the sun god, Ra Hor-akheti, who is represented by the head of a falcon. This statue is distinguished by the fact that it is carved from a unique piece of colored granite stone, which combines the colors red, like the sun, and black, like the fertile ground. The king’s features indicate his narrow eyes and his mouth, which appears naturally to This statue may date back to the end of the thirty-eighth year of the reign of King Amenhotep III."
+    },
     {
       "image": "assets/images/قناع-الملك-توت-عنخ-آمون.jpg",
       "name": "King Tutankhamun",
+      "namee": "الملك توت عنخ امون",
       "details": "Tutankhamun, an Egyptian pharaoh, ascended the throne at a young age, around 9 years old, during the 18th dynasty. His reign was brief, from approximately 1332 to 1323 BCE. He is best known for the intact tomb discovered in the Valley of the Kings in 1922, filled with treasures and artifacts, shedding light on ancient Egyptian culture and history. His death at a young age remains a mystery, sparking speculation and fascination for centuries.",
       "explore": "Tutankhamun, often referred to as King Tut, reigned as the Pharaoh of Egypt from approximately 1332 to 1323 BCE during the New Kingdom's 18th dynasty. Although his reign was relatively short and he ascended to the throne at a very young age, his achievements and the historical significance of his reign are notable. Here are some key achievements and contributions of Tutankhamun:\n\n"
           "1. Religious Restoration: One of Tutankhamun's most significant achievements was the restoration of the traditional Egyptian religion and practices that had been altered by his predecessor, Akhenaten, who promoted the worship of Aten, the sun disk. Tutankhamun, under the influence of his advisors, reinstated the worship of the old gods, particularly Amun, and restored the temples and priesthoods that had fallen into disrepair during Akhenaten's monotheistic reforms.\n\n"
@@ -13,74 +200,7 @@ class PharaohData {
           "6. Tomb Construction: The construction of Tutankhamun’s tomb (KV62) in the Valley of the Kings, though small compared to those of other pharaohs, is one of the best-preserved and most richly furnished tombs discovered. Its treasures have provided invaluable insights into the material culture, craftsmanship, and daily life of ancient Egypt.\n\n"
           "Tutankhamun's death at a young age (around 18 or 19) and the subsequent discovery of his nearly intact tomb in 1922 has greatly contributed to his fame and the understanding of ancient Egyptian civilization. His reign, though brief, played a crucial role in the transitional period of Egyptian history, particularly in the restoration of the traditional religious practices that would continue for centuries."
     },
-    {
-      "image": "assets/images/amnthotb el talt (1).jpg",
-      "name": "King Amnthotb III",
-      "details": "Amenhotep III, also known as Amenhotep the Magnificent, ruled ancient Egypt during the 18th dynasty from around 1386 to 1353 BCE. His reign was marked by prosperity, cultural flourishing, and extensive building projects, including the Temple of Luxor and the Colossi of Memnon. He was known for his diplomatic prowess, maintaining peaceful relations and trade with neighboring kingdoms.",
-      "explore": "Amenhotep III, also known as Amenhotep the Magnificent, ruled over Egypt during the 18th dynasty (circa 1386–1353 BCE) and achieved numerous accomplishments during his reign. Here are some of his notable achievements:\n\n"
-          "1. Building Projects: Amenhotep III was renowned for his extensive building projects, including the construction and renovation of temples, palaces, and monuments throughout Egypt. He built or expanded temples at Karnak, Luxor, and Soleb, among others. His mortuary temple, known as the 'Temple of Millions of Years,' was located on the west bank of the Nile at Thebes and was one of the largest and most impressive structures of its time.\n\n"
-          "2. Art and Culture: Amenhotep III was a patron of the arts and presided over a flourishing period of artistic and cultural achievement in ancient Egypt. He commissioned numerous statues, reliefs, and other works of art, many of which depicted him and his family in idealized forms. The period of his reign is often considered a golden age of Egyptian art.\n\n"
-          "3. Foreign Relations: Amenhotep III maintained peaceful relations with foreign powers during his reign, which contributed to stability and prosperity within Egypt. He engaged in diplomatic marriages with foreign princesses to strengthen alliances, including his marriage to Tiy, who became one of the most influential queens in Egyptian history.\n\n"
-          "4. Economic Prosperity: Under Amenhotep III's rule, Egypt experienced a period of economic prosperity and abundance. Trade flourished, bringing wealth and exotic goods from distant lands, and the agricultural sector thrived, leading to increased food production and prosperity for the population.\n\n"
-          "5. Religious Contributions: Amenhotep III was deeply involved in the religious life of Egypt and sponsored the construction and renovation of temples dedicated to various gods and goddesses. He was particularly devoted to the god Amun-Ra, the principal deity of the Egyptian pantheon, and his religious activities contributed to the power and influence of the priesthood of Amun.\n\n"
-          "6. Monumental Statues: Amenhotep III commissioned numerous colossal statues of himself, including the famous Colossi of Memnon, which stood at the entrance to his mortuary temple. These statues, carved from massive blocks of stone, were intended to showcase the pharaoh's power and divine status.\n\n"
-          "7. Longevity and Stability: Amenhotep III's reign was one of the longest and most stable in ancient Egyptian history, lasting for nearly four decades. His ability to maintain peace, stability, and prosperity throughout his reign contributed to his reputation as one of Egypt's most successful and beloved pharaohs."
-    },
-    {"image": "assets/images/hatshbsot.jpg", "name": "King Hatshbsot",
-      "details":"Hatshepsut, one of ancient Egypt's most renowned female rulers, reigned as pharaoh from around 1479 to 1458 BCE during the 18th dynasty. She famously dressed as a male pharaoh to assert her authority and legitimacy, commissioning ambitious building projects such as the Temple of Hatshepsut at Deir el-Bahari. Under her rule, Egypt experienced peace, prosperity, and extensive trade expeditions to the Land of Punt, enriching the kingdom's resources and cultural exchangeز",
-    "explore":"Hatshepsut, one of ancient Egypt's most remarkable rulers, achieved numerous accomplishments during her reign as pharaoh (circa 1478–1458 BCE). Here are some of her notable achievements:\n\n"
-        "1. First Female Pharaoh: Hatshepsut is perhaps best known for being one of the few women to rule as pharaoh in ancient Egypt. She took the throne after the death of her husband, Thutmose II, and acted as regent for her stepson, Thutmose III, before eventually assuming full pharaonic powers.\n\n"
-        "2. Building Projects: She initiated ambitious building projects throughout Egypt, including the construction of magnificent temples and monuments. The most famous of these is her mortuary temple, Djeser-Djeseru (\"Holy of Holies\"), located at Deir el-Bahri near Thebes. This temple is renowned for its innovative architecture and stunning reliefs.\n\n"
-        "3. Trade and Diplomacy: Hatshepsut's reign saw an expansion of trade networks and diplomatic relations with other civilizations, particularly in the eastern Mediterranean and the land of Punt (possibly modern-day Somalia or Eritrea). Trade missions to Punt brought back exotic goods such as myrrh, ebony, ivory, and live animals.\n\n"
-        "4. Religious Contributions: Hatshepsut played a significant role in the religious life of Egypt, promoting the worship of traditional gods and goddesses while also elevating the status of some lesser-known deities. She associated herself closely with the god Amun, claiming divine lineage and portraying herself as his earthly representative.\n\n"
-        "5. Cultural Patronage: Hatshepsut supported art, literature, and culture during her reign, commissioning statues, reliefs, and other works of art that celebrated her accomplishments and perpetuated her legacy.\n\n"
-        "6. Stability and Prosperity: Under Hatshepsut's rule, Egypt experienced a period of stability and prosperity, with economic growth, trade expansion, and cultural flourishing. Her reign is often remembered as a time of peace and prosperity in ancient Egyptian history.\n\n"
-        "7. Legacy and Succession: Despite efforts to erase her legacy by later pharaohs, including her stepson Thutmose III, Hatshepsut's achievements endured through the centuries. She remains one of ancient Egypt's most celebrated and enigmatic rulers, remembered for her exceptional leadership and the lasting impact of her reign."},
 
-    {"image": "assets/images/akhnton.jpg", "name": "King Akhenaten",
-      "details":"Akhenaten, known for his revolutionary religious reforms, ruled ancient Egypt during the 18th dynasty from around 1353 to 1336 BCE. He introduced monotheistic worship centered on the sun god Aten, shifting away from traditional polytheism. His capital, Akhetaten (modern-day Amarna), reflected his new religious ideology. Despite initial prosperity, his reign faced opposition, and after his death, his reforms were largely reversed, and his monuments were dismantled.",
-      "explore":"Akhenaten, also known as Amenhotep IV, is one of ancient Egypt's most intriguing and controversial pharaohs, primarily known for his religious reforms and radical changes in artistic style. Here are some of his notable achievements:\n\n"
-      "1. Religious Revolution: Akhenaten is most famous for his religious revolution, which involved the elevation of the sun god Aten to the status of the supreme deity and the suppression of traditional Egyptian polytheism. He proclaimed Aten as the one true god and initiated a monotheistic cult centered around the worship of the sun disc. This represented a significant departure from the traditional polytheistic beliefs of ancient Egypt.\n\n"
-          "2. Establishment of Akhetaten: Akhenaten founded a new capital city called Akhetaten (modern-day Amarna) dedicated to the worship of Aten. The city was constructed from scratch and served as the administrative and religious center of Egypt during Akhenaten's reign. It featured grand temples and palaces, reflecting the new religious and artistic ideals of the Aten cult.\n\n"
-          "3. Artistic Innovations: Akhenaten's reign witnessed a revolution in artistic style known as the \"Amarna art.\" This art style deviated from traditional Egyptian artistic conventions, depicting the pharaoh and his family in a more naturalistic and intimate manner. It emphasized realism, with elongated proportions, exaggerated features, and a focus on the royal family's everyday activities.\n\n"
-          "4. Royal Family: Akhenaten's family played a prominent role in his religious reforms and artistic innovations. His wife, Nefertiti, was a powerful and influential queen who played a central role in the Aten cult. Their daughters, especially the famous Ankhesenamun, were also prominent figures in Akhenaten's court.\n\n"
-          "5. Diplomacy and Foreign Policy: Akhenaten's reign saw a shift in Egypt's foreign policy, with a focus on diplomacy and pacifism rather than military conquest. He sought to maintain peaceful relations with neighboring powers and engaged in diplomatic correspondence with other rulers, including those in the Hittite Empire.\n\n"
-          "6. Literary Contributions: Akhenaten is credited with composing a series of hymns and prayers dedicated to the Aten, known as the \"Great Hymn to the Aten.\" These texts praise the sun god and express Akhenaten's devotion to Aten and his role as the sole mediator between Aten and humanity.\n\n"
-          "7. Legacy and Controversy: Despite his significant achievements, Akhenaten's reign ended in controversy, and his religious reforms were largely dismantled by his successors. Nevertheless, his reign represents a fascinating period of experimentation and innovation in ancient Egyptian history, leaving a lasting impact on art, religion, and culture in the Nile Valley."},
-
-   {"image": "assets/images/sonsrt3.jpg", "name": "King Sonsrt III",
-     "details":"Senusret III, a pharaoh of the Middle Kingdom's 12th dynasty, ruled ancient Egypt from approximately 1878 to 1839 BCE. He is celebrated for his military campaigns, expanding Egypt's borders and securing its territories. Senusret III initiated numerous construction projects, including the famous White Chapel at Karnak. His reign is regarded as a period of stability and prosperity, often considered one of the peaks of Egyptian civilization.",
-     "explore":"Senusret III, also known as Sesostris III, was a powerful pharaoh of the Middle Kingdom of Egypt who ruled during the 12th Dynasty (circa 1878–1839 BCE). He is often regarded as one of the most capable and successful rulers in ancient Egyptian history. Here are some of his notable achievements:\n\n"
-     "1. Military Campaigns: Senusret III conducted numerous military campaigns to expand Egypt's territory and maintain control over its borders. He waged successful campaigns against Nubia (to the south), the Levant (to the northeast), and Libya (to the west), securing Egypt's dominance in the region.\n\n"
-         "2. Fortification Projects: Senusret III undertook extensive fortification projects along Egypt's borders to protect the kingdom from external threats. He built a series of fortresses, forts, and military outposts along strategic points on the Nile and in key regions such as the Sinai Peninsula and Lower Nubia.\n\n"
-         "3. Infrastructure Development: Senusret III initiated significant infrastructure projects aimed at improving transportation, communication, and irrigation systems throughout Egypt. He constructed or renovated canals, dikes, and waterways to control the annual flooding of the Nile and ensure agricultural prosperity.\n\n"
-         "4. Monumental Architecture: Senusret III commissioned several impressive architectural projects, including temples, palaces, and statues dedicated to the gods and to commemorate his reign. Notable examples include the renovation of the Temple of Karnak in Thebes and the construction of the White Chapel, a small temple made of white limestone adorned with reliefs depicting the king's divine status.\n\n"
-         "5. Administrative Reforms: Senusret III implemented administrative reforms aimed at centralizing royal authority and strengthening the bureaucratic apparatus of the Egyptian state. He appointed trusted officials to key administrative positions and established systems for taxation, law enforcement, and land management.\n\n"
-         "6. Cultural Patronage: Senusret III was a patron of the arts and literature, supporting the production of religious texts, hymns, and literary works that celebrated his achievements and glorified the gods. His reign witnessed a flourishing of intellectual and artistic activity, with scholars and artisans producing works of exceptional quality and beauty.\n\n"
-         "7. Longevity and Stability: Senusret III's reign was characterized by stability, prosperity, and relative peace, allowing Egypt to thrive economically and culturally. His efficient governance and military prowess contributed to the longevity of the 12th Dynasty and secured his legacy as one of ancient Egypt's most successful pharaohs."},
-
-    {"image": "assets/images/sonsrt1.jpg", "name": "King Sonsrt I",""
-        "details":"Senusret I, a pharaoh of the Middle Kingdom's 12th dynasty, ruled ancient Egypt from approximately 1971 to 1926 BCE. He is renowned for his military campaigns, notably against the Nubians, expanding Egypt's territory and influence. Senusret I oversaw numerous construction projects, including the famous White Chapel and the Pyramid of El-Lahun. His reign is remembered as a period of prosperity and cultural flourishing, marking one of the heights of the Middle Kingdom.",
-      "explore":"Senusret I, also known as Senwosret I, was a significant pharaoh of the Middle Kingdom's 12th Dynasty in ancient Egypt, ruling from around 1971 to 1926 BCE. He is often regarded as one of the most capable and influential rulers of his dynasty. Here are some of his notable achievements:\n\n"
-      "1. Building Projects: Senusret I initiated several ambitious building projects throughout Egypt, including the construction and renovation of temples, palaces, and monuments. He built temples dedicated to various gods and goddesses, such as the Temple of Amun at Karnak and the Temple of Montu at Tod.\n\n"
-          "2. Expansion of Irrigation: Senusret I is credited with expanding Egypt's irrigation systems, particularly in the Fayyum region. He commissioned the construction of canals and waterways to improve agricultural productivity and increase the arable land available for cultivation, leading to economic prosperity.\n\n"
-          "3. Military Campaigns: Senusret I conducted successful military campaigns to defend Egypt's borders and expand its territory. He waged campaigns against the Nubians in the south and the Libyans in the west, securing Egypt's dominance over these regions and ensuring stability along the borders.\n\n"
-          "4. Monumental Statuary: Senusret I commissioned numerous statues and reliefs depicting himself in various poses and royal regalia. These statues served not only to glorify the king but also to propagate royal ideology and affirm his divine kingship to the Egyptian people.\n\n"
-          "5. Cultural Patronage: Senusret I was a patron of the arts and literature, supporting the production of religious texts, hymns, and literary works that celebrated the achievements of the pharaoh and extolled the virtues of the gods. His reign witnessed a flourishing of intellectual and artistic activity, with scholars and artisans producing works of exceptional quality.\n\n"
-          "6. Strengthening of Central Authority: Senusret I implemented administrative reforms aimed at centralizing royal authority and strengthening the bureaucratic apparatus of the Egyptian state. He appointed trusted officials to key administrative positions and established systems for taxation, law enforcement, and land management.\n\n"
-          "7. Legacy and Succession: Senusret I's reign marked a period of stability and prosperity in ancient Egypt, laying the foundation for the subsequent glory of the 12th Dynasty. His efficient governance and military prowess secured his legacy as one of Egypt's most successful and revered pharaohs, revered by later generations for his accomplishments and contributions to the kingdom."},
-
-    {"image": "assets/images/to7tmos.jpg", "name": "King Tahtmos III",
-      "details":"Thutmose III, often referred to as the Napoleon of Egypt, reigned during the New Kingdom's 18th dynasty, from about 1479 to 1425 BCE. He expanded Egypt's empire through numerous military campaigns, including the famous Battle of Megiddo. Thutmose III implemented administrative reforms and encouraged trade, leading to a period of prosperity and cultural advancement known as the Thutmoside Renaissance His reign solidified Egypt's status as a dominant power in the ancient Near East.",
-    "explore":"Thutmose III, often referred to as the Napoleon of Egypt was one of ancient Egypt's most successful warrior pharaohs, reigning from approximately 1479 to 1425 BCE during the New Kingdom period. His achievements are numerous and significant, shaping the course of Egyptian history. Here are some of his notable accomplishments:\n\n"
-    "1. Military Campaigns and Expansion: Thutmose III conducted numerous military campaigns to expand Egypt's borders and establish its dominance in the ancient Near East. He led at least 17 military campaigns into Syria, Canaan, and Nubia, extending Egypt's territory to its greatest extent.\n\n"
-        "2. Battle of Megiddo: Thutmose III's most famous military victory occurred at the Battle of Megiddo, where he defeated a coalition of Canaanite and other regional powers. This decisive victory is one of the earliest recorded battles in history and solidified Egypt's control over the Levant.\n\n"
-        "3. Administrative Reforms: Thutmose III implemented administrative reforms aimed at centralizing royal authority and streamlining the bureaucracy. He appointed trusted officials to key administrative positions and established systems for taxation, law enforcement, and land management.\n\n"
-        "4. Monumental Building Projects: Thutmose III undertook extensive building projects, including the construction and renovation of temples, palaces, and monuments throughout Egypt. Notable examples include the expansion of the Karnak Temple complex and the construction of the Festival Hall at Karnak.\n\n"
-        "5. Cultural Patronage: Thutmose III was a patron of the arts and literature, supporting the production of religious texts, hymns, and artistic works that celebrated his military victories and glorified the gods. He commissioned statues and reliefs depicting himself in various poses and royal regalia.\n\n"
-        "6. Diplomacy and Trade: Thutmose III engaged in diplomacy and trade with foreign powers, forging alliances and establishing commercial relations with neighboring kingdoms. He negotiated treaties and marriage alliances to secure peace and strengthen Egypt's position in the region.\n\n"
-        "7. Temple Restorations: Thutmose III restored and renovated many temples across Egypt, dedicating resources to the upkeep and maintenance of religious structures. He also commissioned new temples and monuments to honor the gods and commemorate his military victories.\n\n"
-        "8. Longevity and Stability: Thutmose III's reign marked a period of stability and prosperity in ancient Egypt, known as the 'Golden Age' of the New Kingdom. His military conquests and administrative reforms contributed to Egypt's wealth and power, ensuring his legacy as one of Egypt's greatest pharaohs."},
   ];
 
   String getImage(int index) {
@@ -95,5 +215,9 @@ class PharaohData {
   }
   String getExplore(int index) {
     return pharaoh[index]['explore']!;
+  }
+
+  String getNamee(int index) {
+    return pharaoh[index]['namee']!;
   }
 }
