@@ -1,15 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:pharera/pharahosl_ist.dart';
+import 'fav_but.dart';
+import 'pharahos_list.dart';
+
+List<String> favoriteImages = [];
 
 class PharaohDetailPage extends StatelessWidget {
   final PharaohData pharaohData;
   final int index;
 
-  const PharaohDetailPage({Key? key, required this.pharaohData, required this.index})
-      : super(key: key);
+  const PharaohDetailPage({
+    Key? key,
+    required this.pharaohData,
+    required this.index,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    String imagePath = pharaohData.getImage(index);
+
     return Scaffold(
       backgroundColor: const Color.fromARGB(255, 226, 226, 226),
       appBar: AppBar(
@@ -26,27 +34,34 @@ class PharaohDetailPage extends StatelessWidget {
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
             Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20.0, vertical: 20.0),
-              child: Container(
-                width: 400,
-                height: 430,
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(20),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.black.withOpacity(0.5),
-                      spreadRadius: 5,
-                      blurRadius: 7,
-                      offset: const Offset(0, 3),
-                    ),
-                  ],
-                ),
-                child: ClipRRect(
-                  borderRadius: BorderRadius.circular(20),
-                  child: Image.asset(
-                    pharaohData.getImage(index),
-                    fit: BoxFit.cover,
-                    height: 750,
+              padding: const EdgeInsets.symmetric(horizontal: 50.0, vertical: 20.0),
+              child: Center(
+                child: Container(
+                  width: 230,
+                  height: 430,
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(20),
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.black.withOpacity(0.5),
+                        spreadRadius: 5,
+                        blurRadius: 7,
+                        offset: const Offset(0, 3),
+                      ),
+                    ],
+                  ),
+                  child: Stack(
+                    children: [
+                      ClipRRect(
+                        borderRadius: BorderRadius.circular(20),
+                        child: Image.asset(
+                          imagePath,
+                          width: 250, // تحديد عرض الصورة
+                          height: 430, // تحديد ارتفاع الصورة
+                          fit: BoxFit.fill,
+                        ),
+                      ),
+                    ],
                   ),
                 ),
               ),
@@ -89,7 +104,7 @@ class PharaohDetailPage extends StatelessWidget {
                           ),
                           const SizedBox(height: 20),
                           Text(
-                            pharaohData.getExplore(index),
+                            pharaohData.getDetails(index),
                             style: const TextStyle(
                               fontSize: 18,
                               color: Colors.black,
